+++ docs/versionhistory.rst
@@ -1,6 +1,24 @@
 Version history
 ===============
 
+**2.0.0**
+
+- *BACKWARD INCOMPATIBLE* Dropped Python 3.4 support in order to make the code fully rely on the
+  new ``async``/``await``, ``async for`` and ``async with`` language additions
+- *BACKWARD INCOMPATIBLE* De-emphasized the ability to run code in worker threads.
+  It is now recommended to minimize the use of worker threads.
+  As a result, the ``@asynchronous`` decorator has been removed.
+- *BACKWARD INCOMPATIBLE* Replaced the ``@blocking`` decorator with the ``threadpool`` function
+  that acts both as a decorator and an asynchronous context manager that executes wrapped code in
+  a worker thread
+- *BACKWARD INCOMPATIBLE* Removed the deprecated ``blocking`` and ``asynchronous`` aliases from the
+  ``asphalt.core.util`` module
+- *BACKWARD INCOMPATIBLE* Removed regular context manager support from the ``Context`` class
+  (asynchronous context manager support still remains)
+- Added an asynchronous wrapper for running file I/O operations in a worker thread
+- Added wrappers for easily creating asynchronous generators and context managers
+
+
 **1.3.0**
 
 - Allow the context manager of the ``Context`` class to be used from a non-eventloop thread when